refactor(tooling): split dev app support
